Instructli (Typescrit + Django + React)/Autumn2025ComputerOrganizationAITutor/backend/requirements.txt
dotenv==0.9.9
r2r==3.6.5
django==5.2.3
djangorestframework==3.16.0
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.0
django-filter==24.3
social-auth-app-django==5.5.1
gunicorn==23.0.0
django-redis==6.0.0
uuid==1.30
openai==1.99.5
psycopg2-binary==2.9.10